show related dataset title instead of slug

+++ partials/app-entry.php
@@ -27,16 +27,21 @@ if ( $icon_attributes ) {
 		echo '</p>';
 		echo '<p>' . the_content() . '</p>';
 		if ( ! empty( $related_datasets ) ) {
-			echo '<div class="collapse" id="collapsed-related-' . esc_attr( get_the_ID() ) . '">';
-			echo '<ul>';
-			foreach ( $related_datasets as $dataset ) {
-				echo '<li><a href="' . esc_attr( get_page_link_by_slug( 'dataset/' . $dataset['dataset_id'] ) ) . '">' . esc_attr( $dataset['dataset_id'] ) . '</a></li>';
+			// Check if wp-ckan-backend plugin is active
+			if ( ! method_exists( 'Ckan_Backend_Helper', 'get_dataset_title' ) ) {
+				esc_attr_e( 'Please activate wp-ckan-backend plugin', 'ogdch' );
+			} else {
+				echo '<div class="collapse" id="collapsed-related-' . esc_attr( get_the_ID() ) . '">';
+				echo '<ul>';
+				foreach ( $related_datasets as $dataset ) {
+					$title = Ckan_Backend_Helper::get_dataset_title( $dataset['dataset_id'] );
+					echo '<li><a href="' . esc_attr( get_page_link_by_slug( 'dataset/' . $dataset['dataset_id'] ) ) . '">' . esc_attr( $title ) . '</a></li>';
+				}
+				echo '</ul>';
+				echo '</div>';
+				echo '<p class="small"><a id="collapsed-related-' . esc_attr( get_the_ID() ) . '-link" data-toggle="collapse" href="#collapsed-related-' . esc_attr( get_the_ID() ) . '" aria-expanded="false" aria-controls="collapsed-related-' . esc_attr( get_the_ID() ) . '">' . esc_attr__( 'Show related datasets', 'ogdch' ) . '</a></p>';
 			}
-			echo '</ul>';
-			echo '</div>';
-			echo '<p class="small"><a id="collapsed-related-' . esc_attr( get_the_ID() ) . '-link" data-toggle="collapse" href="#collapsed-related-' . esc_attr( get_the_ID() ) . '" aria-expanded="false" aria-controls="collapsed-related-' . esc_attr( get_the_ID() ) . '">' . esc_attr__( 'Show related datasets', 'ogdch' ) . '</a></p>';
 		}
-
 		?>
 	</div>
 </div>
